+++ src/main/java/de/markusbordihn/cats/data/HappinessSource.java
@@ -25,7 +25,8 @@ public enum HappinessSource {
   PETTING(5, PersonalityType.CUDDLY),
   PLAYING(8, PersonalityType.PLAYFUL),
   TREAT(10, PersonalityType.LAZY),
-  FEEDING(3, null);
+  FEEDING(3, null),
+  SLEEPING_IN_BED(10, PersonalityType.LAZY);
 
   private final int baseDelta;
   @Nullable private final PersonalityType preferredPersonality;
@@ -46,7 +47,7 @@ public enum HappinessSource {
         switch (this) {
           case PETTING -> catPersonality.getPetModifier();
           case PLAYING -> catPersonality.getPlayModifier();
-          case TREAT, FEEDING -> catPersonality.getActivityModifier();
+          case TREAT, FEEDING, SLEEPING_IN_BED -> catPersonality.getActivityModifier();
         };
     return Math.max(1, (int) (baseDelta * modifier));
   }