+++ src/main/java/de/markusbordihn/cats/interaction/InteractionWild.java
@@ -29,8 +29,6 @@ public class InteractionWild {
 
   public static boolean handle(
       Ref<EntityStore> entityRef, Role role, Store<EntityStore> store, Player player) {
-    InteractionLogger.logInteraction(
-        "WILD CAT: Basic Interaction", entityRef, role, store, player, null);
     return false;
   }
 }